requirements.txt
django
Pillow
pandas
matplotlib
openai
google-generativeai
huggingface_hub
python-dotenv
numpy
requests